refactor user model

+++ src/app/models/User.js
@@ -8,7 +8,7 @@ class User extends Model {
       {
         name: Sequelize.STRING,
         email: Sequelize.STRING,
-        cellphone: Sequelize.BIGINT,
+        cellphone: Sequelize.INTEGER,
         password: Sequelize.STRING,
       },
       {
